pyproject.toml
[project]
name = "trex-effect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "pygame-ce>=2.5.6",
]
